enforce types across front and backend
